+++ docs/whatsnew.rst
@@ -8,8 +8,20 @@ The full release history can be viewed `at the GitHub yank releases page <https:
 
 0.25.0 Current development
 --------------------------
+
+API-breaking changes
+^^^^^^^^^^^^^^^^^^^^
 - The ``yank.mpi`` module and the objects in the ``yank.multistate`` package, which were deprecated in 0.24.0, have now
   been removed, and they can be found in the ``mpiplus`` and ``openmmtools`` libraries respectively.
+
+New features
+^^^^^^^^^^^^
+- The trailblaze algorithm used for the authomatic generation of the alchemical path is now capable of resuming after an unexpected interruption or crash. The samples generated during the process are used to initialize the replicas of the replica exchange or SAMS free energy calculation (`#1176 <https://github.com/choderalab/yank/pull/1176>`_).
+
+Enhancements
+^^^^^^^^^^^^
+- By default, the automatic determination of the alchemical path now starts with the harmonic/flat-bottom restraint turned off and activate it in intermediate states instead of keeping the restraint activated throughout the calculation and reweighting in the analysis stage (`#1176 <https://github.com/choderalab/yank/pull/1176>`_).
+
 
 0.24.1 Bugfix release
 ---------------------